requirements.txt
fastapi
uvicorn
pydantic
pydantic-settings
firecrawl-py
tenacity
python-dotenv
requests
beautifulsoup4
google-cloud-bigquery
db-dtypes
python-json-logger
